+++ common/src/main/java/io/github/mortuusars/scholar/Config.java
@@ -16,7 +16,8 @@ public class Config {
         public static final ForgeConfigSpec.BooleanValue LECTERN_TWO_PAGE_BOOK_SCREEN;
         public static final ForgeConfigSpec.BooleanValue SNEAK_OPENS_VANILLA_BOOK_SCREEN;
 
-        // QOL
+        // Misc
+        public static final ForgeConfigSpec.BooleanValue CHISELED_BOOKSHELF_COLORS;
         public static final ForgeConfigSpec.BooleanValue CHISELED_BOOKSHELF_TOOLTIP;
         public static final ForgeConfigSpec.BooleanValue BOOK_ENCHANTMENT_GLINT;
 
@@ -46,6 +47,12 @@ public class Config {
             builder.pop();
 
             builder.push("misc");
+            CHISELED_BOOKSHELF_COLORS = builder
+                    .comment("Colored books in Chiseled Bookshelf will have correct colors displayed on the block. Default: true",
+                            "Note: resourepacks that modify Chiseled Bookshelf may break the coloring.",
+                            "Note 2: even if this setting is disabled - bookshelf will not look quite the same if you look closely.",
+                            "To restore fully - overwrite bookshelf slot models added by Scholar using a resourcepack.")
+                    .define("chiseled_bookshelf_colors", true);
             CHISELED_BOOKSHELF_TOOLTIP = builder
                     .comment("Hovering over a slot in a Chiseled Bookshelf will show tooltip of a book that's stored in that slot. Default: true")
                     .define("chiseled_bookshelf_tooltip", true);