+++ common/src/main/java/io/github/mortuusars/scholar/mixin/chiseled_bookshelf/ChiseledBookShelfBlockEntityMixin.java
@@ -1,0 +1,41 @@
+package io.github.mortuusars.scholar.mixin.chiseled_bookshelf;
+
+import io.github.mortuusars.scholar.Config;
+import net.minecraft.core.NonNullList;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.network.protocol.Packet;
+import net.minecraft.network.protocol.game.ClientGamePacketListener;
+import net.minecraft.network.protocol.game.ClientboundBlockEntityDataPacket;
+import net.minecraft.world.ContainerHelper;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.block.entity.ChiseledBookShelfBlockEntity;
+import org.jetbrains.annotations.Nullable;
+import org.spongepowered.asm.mixin.Final;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+
+/**
+ * Sends stored items to client, so it can tint book properly and display bookshelf contents in a tooltip.
+ */
+@Mixin(ChiseledBookShelfBlockEntity.class)
+public abstract class ChiseledBookShelfBlockEntityMixin extends BlockEntityParentMixin {
+    @Shadow
+    @Final
+    private NonNullList<ItemStack> items;
+
+    @Override
+    protected Packet<ClientGamePacketListener> onGetUpdatePacket(@Nullable Packet<ClientGamePacketListener> packet) {
+        return Config.Common.CHISELED_BOOKSHELF_TOOLTIP.get() || Config.Common.CHISELED_BOOKSHELF_COLORS.get()
+                ? ClientboundBlockEntityDataPacket.create((ChiseledBookShelfBlockEntity) (Object) this)
+                : super.onGetUpdatePacket(packet);
+    }
+
+    @Override
+    protected CompoundTag onGetUpdateTag(CompoundTag tag) {
+        if (Config.Common.CHISELED_BOOKSHELF_TOOLTIP.get() || Config.Common.CHISELED_BOOKSHELF_COLORS.get()) {
+            ContainerHelper.saveAllItems(tag, this.items);
+        }
+
+        return tag;
+    }
+}