+++ fabric/src/main/java/io/github/mortuusars/scholar/fabric/ScholarFabricClient.java
@@ -5,13 +5,11 @@ import io.github.mortuusars.scholar.ScholarClient;
 import io.github.mortuusars.scholar.book.BookColor;
 import io.github.mortuusars.scholar.client.gui.screen.edit.LecternSpreadBookEditScreen;
 import io.github.mortuusars.scholar.client.gui.screen.view.LecternSpreadBookViewScreen;
-import io.github.mortuusars.scholar.client.render.ChiseledBookShelfOverlay;
+import io.github.mortuusars.scholar.client.render.ChiseledBookShelf;
 import io.github.mortuusars.scholar.network.fabric.PacketsImpl;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.blockrenderlayer.v1.BlockRenderLayerMap;
 import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper;
-import net.fabricmc.fabric.api.client.model.loading.v1.ModelLoadingPlugin;
-import net.fabricmc.fabric.api.client.model.loading.v1.ModelModifier;
 import net.fabricmc.fabric.api.client.rendering.v1.ColorProviderRegistry;
 import net.fabricmc.fabric.api.client.rendering.v1.HudRenderCallback;
 import net.minecraft.client.gui.screens.MenuScreens;
@@ -23,23 +21,23 @@ public class ScholarFabricClient implements ClientModInitializer {
     @Override
     public void onInitializeClient() {
         BlockRenderLayerMap.INSTANCE.putBlock(Blocks.CHISELED_BOOKSHELF, RenderType.cutout());
-        ColorProviderRegistry.BLOCK.register(BookColor::getBlockTintColor, Blocks.CHISELED_BOOKSHELF);
+        ColorProviderRegistry.BLOCK.register(ChiseledBookShelf::getSlotTintColor, Blocks.CHISELED_BOOKSHELF);
 
-        ColorProviderRegistry.ITEM.register(BookColor::getTintColor, Items.WRITABLE_BOOK);
-        ColorProviderRegistry.ITEM.register(BookColor::getTintColor, Items.WRITTEN_BOOK);
+        ColorProviderRegistry.ITEM.register(BookColor::getItemTintColor, Items.WRITABLE_BOOK);
+        ColorProviderRegistry.ITEM.register(BookColor::getItemTintColor, Items.WRITTEN_BOOK);
 
         ScholarClient.KeyMappings.register(KeyBindingHelper::registerKeyBinding);
 
         MenuScreens.register(Scholar.MenuTypes.LECTERN_SPREAD_BOOK_VIEW.get(), LecternSpreadBookViewScreen::new);
         MenuScreens.register(Scholar.MenuTypes.LECTERN_SPREAD_BOOK_EDIT.get(), LecternSpreadBookEditScreen::new);
 
-        HudRenderCallback.EVENT.register(ChiseledBookShelfOverlay::render);
+        HudRenderCallback.EVENT.register(ChiseledBookShelf::renderSlotTooltip);
 
         PacketsImpl.registerS2CPackets();
 
-        ModelLoadingPlugin.register((ModelLoadingPlugin.Context context) -> {
-            context.modifyModelAfterBake()
-                    .register(ModelModifier.WRAP_PHASE, ChiseledBookshelfBakedModelFabric::modifyModelAfterBake);
-        });
+//        ModelLoadingPlugin.register((ModelLoadingPlugin.Context context) -> {
+//            context.modifyModelAfterBake()
+//                    .register(ModelModifier.WRAP_PHASE, ChiseledBookshelfBakedModelFabric::modifyModelAfterBake);
+//        });
     }
 }