+++ forge/src/main/java/io/github/mortuusars/scholar/forge/event/ClientEvents.java
@@ -5,7 +5,7 @@ import io.github.mortuusars.scholar.ScholarClient;
 import io.github.mortuusars.scholar.book.BookColor;
 import io.github.mortuusars.scholar.client.gui.screen.edit.LecternSpreadBookEditScreen;
 import io.github.mortuusars.scholar.client.gui.screen.view.LecternSpreadBookViewScreen;
-import io.github.mortuusars.scholar.client.render.ChiseledBookShelfOverlay;
+import io.github.mortuusars.scholar.client.render.ChiseledBookShelf;
 import net.minecraft.client.gui.screens.MenuScreens;
 import net.minecraft.client.renderer.ItemBlockRenderTypes;
 import net.minecraft.client.renderer.RenderType;
@@ -35,13 +35,13 @@ public class ClientEvents {
 
         @SubscribeEvent
         public static void registerBlockColors(RegisterColorHandlersEvent.Block event) {
-            event.register(BookColor::getBlockTintColor, Blocks.CHISELED_BOOKSHELF);
+            event.register(ChiseledBookShelf::getSlotTintColor, Blocks.CHISELED_BOOKSHELF);
         }
 
         @SubscribeEvent
         public static void registerItemColors(RegisterColorHandlersEvent.Item event) {
-            event.register(BookColor::getTintColor, Items.WRITABLE_BOOK);
-            event.register(BookColor::getTintColor, Items.WRITTEN_BOOK);
+            event.register(BookColor::getItemTintColor, Items.WRITABLE_BOOK);
+            event.register(BookColor::getItemTintColor, Items.WRITTEN_BOOK);
         }
 
         @SubscribeEvent
@@ -54,7 +54,7 @@ public class ClientEvents {
     public static class ForgeBus {
         @SubscribeEvent
         public static void onRenderGuiPost(RenderGuiEvent.Post event) {
-            ChiseledBookShelfOverlay.render(event.getGuiGraphics(), event.getPartialTick());
+            ChiseledBookShelf.renderSlotTooltip(event.getGuiGraphics(), event.getPartialTick());
         }
     }
 }